.. autogit ..

+++ includes/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: bvalette <bvalette@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/13 09:44:34 by bvalette          #+#    #+#             */
-/*   Updated: 2019/12/25 23:25:17 by bvalette         ###   ########.fr       */
+/*   Updated: 2019/12/27 09:22:03 by bvalette         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,7 +88,7 @@ size_t				ft_strlcpy(char *dst, const char *src, size_t dstsize);
 size_t				ft_strlcat(char *dst, const char *src, size_t dstsize);
 char				*ft_strnstr(const char *haystack, const char *needle,
 	size_t len);
-long long int		ft_atoi(const char *str);
+int					ft_atoi(const char *str);
 char				*ft_strdup(const char *s1);
 
 /*
